% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMigrationEvents.R
\name{classifyMigrationEvents}
\alias{classifyMigrationEvents}
\title{Classify Migration Events}
\usage{
classifyMigrationEvents(temp_data, focal_year, allowable_overlap)
}
\arguments{
\item{temp_data}{A data.frame or data.table containing potential migration events, including
columns such as \code{new_start_year}, \code{new_end_year}, \code{old_start_year},
\code{old_end_year}, \code{old_geo}, and \code{new_geo}.}

\item{focal_year}{An integer representing the year in which migration
classification is to be assessed.}

\item{allowable_overlap}{An integer representing the maximum allowable
overlap between the new and old end years for a migration to be classified
as a rupture.}
}
\value{
A data.table containing classified migration events, with an indicator
for migration and the year of migration.
}
\description{
This function classifies migration events based on the presence of
ruptures between new and old affiliations. A rupture is defined
by the overlap of years, while a gap is defined by the time between
old and new affiliations. For more details see Herman, A. (Forthcoming)
"Reconstructing bibliometric methods for studying mobility."
}
\details{
This function is meant for internal use, as it estimates migration events
for a single focal year.
}
