% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMigrationEvents.R
\name{classifyTemporaryEvents}
\alias{classifyTemporaryEvents}
\title{Classify Temporary Events}
\usage{
classifyTemporaryEvents(temp_data, focal_year, allowable_temporary_overlap)
}
\arguments{
\item{temp_data}{A data.frame or data.table containing potential temporary events, including
columns such as \code{new_start_year}, \code{new_end_year}, and \code{old_end_year}.}

\item{focal_year}{An integer representing the year in which temporary
event classification is to be assessed.}

\item{allowable_temporary_overlap}{An integer representing the maximum
allowable duration for a temporary event to be classified as such.}
}
\value{
A data frame containing classified temporary events, with an indicator
for temporary status, the start year of the temporary event, and its length.
}
\description{
This function classifies temporary events based on the relationship
between new and old affiliations, as defined in Herman, A. (Forthcoming)
"Reconstructing bibliometric methods for studying mobility.". A temporary event is
defined as a situation where the new affiliation ends before the old
affiliation ends.
}
