% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAffiliationTenures.R
\name{getAffiliationTenures}
\alias{getAffiliationTenures}
\title{Get Affiliation Tenures}
\usage{
getAffiliationTenures(aff_data, person_var, time_var, geo_var)
}
\arguments{
\item{aff_data}{A data.frame or data.table containing affiliation data with columns for
individual identifiers, time indicators, geographical identifiers, and a backfilled status.}

\item{person_var}{A string specifying the name of the column representing
individual identifiers in \code{aff_data}.}

\item{time_var}{A string specifying the name of the column representing
time indicators in \code{aff_data}.}

\item{geo_var}{A string specifying the name of the column representing
geographical identifiers in \code{aff_data}.}
}
\value{
A data.table with columns for person ID, affiliation or geography ID,
estimated start and end years of tenures, and a boolean indicating whether the
tenure estimate relied on data that was backfilled.
}
\description{
This function processes affiliation data to extract tenure information for individuals,
identifying periods of association with affiliations or geographical locations and creating
unique event IDs.
}
\details{
The function organizes the data by person and geography, identifying the start and end
years of affiliations. It does not return tenures for people with only one affiliation over time.

Unique event IDs are created according to the start date of tenures, the only point at which
there can be tenure events. The event ID is a combination of the start date of tenures, as well
as the tenure IDs of any tenures that started at that point in time.
}
