% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAmbiguousEvents.R
\name{getAmbiguousEvents}
\alias{getAmbiguousEvents}
\title{Get Ambiguous Events}
\usage{
getAmbiguousEvents(migration_data)
}
\arguments{
\item{migration_data}{A data.frame or data.table containing migration data with at least
\code{tenure_event_id}, \code{old_tenure_id}, and \code{new_tenure_id} columns.}
}
\value{
A data.table that includes the original migration data along with two new
logical columns: \code{ambiguous_senders} indicating if a tenure event has multiple senders,
and \code{ambiguous_receivers} indicating if a tenure event has multiple unique
receivers.
}
\description{
This function analyzes migration data to identify ambiguous senders and receivers
based on tenure event identifiers, as defined in Herman, A. (Forthcoming) "Reconstructing
bibliometric methods for studying mobility." It determines whether there are multiple unique
tenure IDs associated with each sender and receiver for a given migration event.
}
