% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensoredEvents.R
\name{getCensoredEvents}
\alias{getCensoredEvents}
\title{Get Censored Events}
\usage{
getCensoredEvents(tenure_data, backfill_range, final_t)
}
\arguments{
\item{tenure_data}{A data.frame or data.table containing tenure information with columns \code{person_id},
\code{est_start_year}, \code{est_end_year}, and \code{backfilled}.}

\item{backfill_range}{An integer indicating the backfill range. This is needed to determine the final
observation year.}

\item{final_t}{An integer representing the final year to assess for ongoing affiliations.}
}
\value{
A list containing two data.tables:
\item{censored_by_last_obs}{A data.table of events that were censored based on the last observation.}
\item{censored_by_time_series}{A data.table of events that were censored based on the time series, specifically those
that have their \code{est_end_year} equal to \code{final_t}.}
}
\description{
This function identifies censored events in tenure data based on specified criteria,
as defined in Herman, A. (Forthcoming) "Reconstructing bibliometric methods for studying mobility."
It returns two event subsets: those censored by the last observation of a person's publication career
and those censored by the final observation in the overall time series.
}
