% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMultipleAffiliations.R
\name{getMultipleAffiliations}
\alias{getMultipleAffiliations}
\title{Get Multiple Affiliations}
\usage{
getMultipleAffiliations(
  tenure_data,
  min_year_of_interest,
  max_year_of_interest
)
}
\arguments{
\item{tenure_data}{A data.frame or data.table containing tenure information with columns for
individual identifiers, geographical identifiers, start and end years of tenures,
and additional related fields.}

\item{min_year_of_interest}{An integer representing the minimum year to
consider for identifying multiple affiliations.}

\item{max_year_of_interest}{An integer representing the maximum year to
consider for identifying multiple affiliations.}
}
\value{
A list containing:
\describe{
\item{ambiguous_simultaneous_starts}{A data frame of ambiguous simultaneous
starts in affiliations.}
\item{multi_tenures}{A data frame of tenure data for individuals with multiple
affiliations.}
\item{multi_tenures_periods}{A data frame of periods when individuals had
multiple affiliations.}
}
}
\description{
This function identifies individuals with multiple affiliations within a specified
range of years based on their tenure data. It checks for ambiguous simultaneous
starts (potentially censored migrants) and determines periods of multiple affiliations.
See Herman, A. (Forthcoming) "Reconstructing bibliometric methods for studying mobility"
for more information.
}
