% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonMobile.R
\name{getNonMobileTenures}
\alias{getNonMobileTenures}
\title{Get Non-Mobile Tenures}
\usage{
getNonMobileTenures(tenure_data)
}
\arguments{
\item{tenure_data}{A data.frame or data.table containing tenure information with columns for
individual identifiers, geographical identifiers, start and end years of tenures.}
}
\value{
A data frame of tenures for non-mobile scientists, containing the following columns:
\describe{
\item{person_id}{Identifier for the individual.}
\item{geo_id}{Identifier for the geographical location.}
\item{est_start_year}{The estimated start year of the tenure.}
\item{est_end_year}{The estimated end year of the tenure.}
}
}
\description{
This function identifies non-mobile scientists based on their tenure data.
Non-mobile scientists are defined as those with only one unique start year
and one unique end year across their tenures. The function returns all
tenures associated with these scientists.
}
