% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMigrationEvents.R
\name{identifyPotentialMigrations}
\alias{identifyPotentialMigrations}
\title{Identify Potential Migrations}
\usage{
identifyPotentialMigrations(tenure_data, focal_year)
}
\arguments{
\item{tenure_data}{A data frame containing tenure information with columns
\code{person_id}, \code{est_start_year}, and \code{est_end_year}, among others.}

\item{focal_year}{An integer representing the year in which new affiliations
are to be assessed for potential migration.}
}
\value{
A data.table containing potential migration events, including details
about both pre-existing and new affiliations, as well as the career start and end
for each individual identified as a potential mover.
}
\description{
This function identifies individuals who may have migrated based on their
affiliations in a specified focal year. It looks for individuals who have a new
position in the focal year while having at least one pre-existing affiliation.
}
